requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
